Random Forest Models

This learner provides fitting procedures for elastic net models, using the glmnet package. For details on the fitting procedure, consult the documentation of the glmnet package.

Lrnr_glmnet

Format

R6Class object.

Value

Lrnr_base object with methods for training and prediction.

Fields

...

Additional arguments. Currently unused.