The Scalable Highly Adaptive LASSO

The Highly Adaptive LASSO is an estimation procedure that generates a design matrix consisting of basis functions corresponding to covariates and interactions of covariates and fits LASSO regression to this (usually) very wide matrix, recovering a nonparametric functional form that describes the target prediction function as a composition of subset functions with finite variation norm. This implementation uses the hal9001 R package, which provides both a custom implementation (based on the origami package) of the CV-LASSO as well the standard call to cv.glmnet from the glmnet package.

Lrnr_hal9001

Format

R6Class object.

Value

Learner object with methods for training and prediction. See Lrnr_base for documentation on learners.

Parameters

degrees="degrees"

The highest order of interaction terms for which the basis functions ought to be generated. The default (NULL) corresponds to generating basis functions for the full dimensionality of the input matrix.

fit_type="fit_type"

The specific routine to be called when fitting the LASSO regression in a cross-validated manner. Choosing the glmnet option will result in a call to cv.glmnet while origami will produce a (faster) call to a custom routine based on the origami package.

n_folds="n_folds"

Integer for the number of folds to be used when splitting the data for cross-validation. This defaults to 10 as this is the convention for v-fold cross-validation.

use_min="use_min"

Determines which lambda is selected from cv.glmnet. TRUE corresponds to "lambda.min" and FALSE corresponds to "lambda.1se".

...

Other parameters passed directly to fit_hal. See its documentation for details.

See also

Other Learners: Custom_chain, Lrnr_HarmonicReg, Lrnr_arima, Lrnr_base, Lrnr_bilstm, Lrnr_condensier, Lrnr_cv, Lrnr_define_interactions, Lrnr_expSmooth, Lrnr_glm_fast, Lrnr_glmnet, Lrnr_glm, Lrnr_h2o_grid, Lrnr_independent_binomial, Lrnr_lstm, Lrnr_mean, Lrnr_nnls, Lrnr_optim, Lrnr_pca, Lrnr_pkg_SuperLearner, Lrnr_randomForest, Lrnr_rugarch, Lrnr_sl, Lrnr_solnp_density, Lrnr_solnp, Lrnr_subset_covariates, Lrnr_tsDyn, Lrnr_xgboost, Pipeline, Stack, define_h2o_X, undocumented_learner