SuperLearner Wrapper Interface

This learner provides an interface to the wrapper functions provided by the SuperLearner package. Such wrapper functions are used to draw on learning algorithms provided by a wide variety of R packages, towards the goal of fitting a SuperLearner ensemble model. Consult documentation of the SuperLearner package for details on these wrapper functions.

Lrnr_pkg_SuperLearner

Format

R6Class object.

Value

Lrnr_base object with methods for training and prediction.

Fields

SL_wrapper

A SuperLearner wrapper, used as an interface to algorithms to be fit in a SuperLearner ensemble model.

...

Additional arguments. Currently unused.