Polyspline - multivariate adaptive polynomial spline regression (polymars) and polychotomous regression and multiple classification (polyclass)

This learner provides fitting procedures for an adaptive regression procedure using piecewise linear splines to model the response, using the function polymars or polyclass from the polspline package as appropriate.

Format

R6Class object.

Value

Learner object with methods for training and prediction. See Lrnr_base for documentation on learners.

Parameters

cv

The number of cross-validation folds to be used in evaluating the sequence of fit models. This is only passed to polyclass for binary/categorical outcomes.

...

Other parameters passed to either polymars or polyclass. See their documentation for details.

See also

Super class

sl3::Lrnr_base -> Lrnr_polspline

Methods

Public methods

Lrnr_polspline$new()

Lrnr_polspline$clone()

Inherited methods

sl3::Lrnr_base$assert_trained()

sl3::Lrnr_base$base_chain()

sl3::Lrnr_base$base_predict()

sl3::Lrnr_base$base_train()

sl3::Lrnr_base$chain()

sl3::Lrnr_base$custom_chain()

sl3::Lrnr_base$get_outcome_type()

sl3::Lrnr_base$predict()

sl3::Lrnr_base$predict_fold()

sl3::Lrnr_base$print()

sl3::Lrnr_base$set_train()

sl3::Lrnr_base$subset_covariates()

sl3::Lrnr_base$train()

sl3::Lrnr_base$train_sublearners()

Method new()

Usage

Lrnr_polspline$new(cv = 5, ...)

Method clone()

The objects of this class are cloneable with this method.

Usage

Lrnr_polspline$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.