Univariate GARCH Models

This learner supports autoregressive fractionally integrated moving average and various flavors of generalized autoregressive conditional heteroskedasticity models for univariate time-series.

Format

R6Class object.

Value

Learner object with methods for training and prediction. See Lrnr_base for documentation on learners.

Parameters

variance.model

List containing variance model specification. This includes model, GARCH order, submodel, external regressors and variance tageting. Refer to ugarchspec for more information.

mean.model

List containing the mean model specification. This includes ARMA model, whether the mean should be included, and external regressors among others. Refer to ugarchspec for more information.

distribution.model="norm"

Conditional density to use for the innovations.

start.pars=list()

List of staring parameters for the optimization routine.

fixed.pars=list()

List of parameters which are to be kept fixed during the optimization.

n.ahead=NULL

The forecast horizon.

See also