Support Vector Machines

This learner uses svm from e1071 to fit a support vector machine (SVM).

Lrnr_svm

Format

R6Class object.

Value

Learner object with methods for training and prediction. See Lrnr_base for documentation on learners.

Parameters

scale = TRUE

A logical vector indicating the variables to be scaled. For a detailed description, please consult the documentation for svm.

type = NULL

SVMs can be used as a classification machine, as a regression machine, or for novelty detection. Depending of whether the outcome is a factor or not, the default setting for this argument is "C-classification" or "eps-regression", respectively. This may be overwritten by setting an explicit value. For a full set of options, please consult the documentation for svm.

kernel = "radial"

The kernel used in training and predicting. You might consider changing some of the optional parameters, depending on the kernel type. Options for kernels include: "linear", "polynomial", "radial" (the default), "sigmoid". For a detailed description, please consult the documentation for svm.

fitted = TRUE

Logical indicating whether the fitted values should be computed and included in the model fit object or not (default: TRUE).

probability = FALSE

Logical indicating whether the model should allow for probability predictions (default: FALSE).

...

Other parameters passed to svm. See its documentation for details.

See also