Nonlinear Time Series Analysis

This learner supports various forms of nonlinear autoregression, including additive AR, neural nets, SETAR and LSTAR models, threshold VAR and VECM.

Lrnr_tsDyn

Format

R6Class object.

Value

Lrnr_base object with methods for training and prediction

Fields

learner

Available built-in time series models. Currently available are listed by with availableModels() function.

m

embedding dimension.

d

time delay.

include

Type of deterministic regressors to include.

type

Whether the variable is taken is level, difference or a mix as in the ADF test.

mL

autoregressive order for low regime.

mH

autoregressive order for high regime.

mM

autoregressive order for middle regime.

thDelay

Time delay for the threshold variable.

common

Indicates which elements are common to all regimes.

ML

vector of lags for order for low.

MM

vector of lags for order for middle.

MH

vector of lags for order for high.

nthresh

Threshold of the model.

trim

trimming parameter indicating the minimal percentage of observations in each regime.

sig

significance level for the tests to select the number of regimes.

control

further arguments to be passed as control list to optim.

r

Number of cointegrating relationships.

model

Model to estimate. Choices: VAR/VECM/TAR/MTAR.

I

For VAR only: whether in the VAR variables are to be taken in levels or as a difference.

beta

For VECM only: imposed cointegrating value.

estim

Type of estimator for the VECM (two-step approach or Johansen MLE).

exogen

Inclusion of exogenous variables.

LRinclude

Possibility to include in the long-run relationship and the ECT trend.

commonInter

Whether the deterministic regressors are regime specific.

mTh

combination of variables with same lag order for the transition variable.

gamma

prespecified threshold values.

dummyToBothRegimes

Whether the dummy in the one threshold model is applied to each regime.

max.iter

Number of iterations for the algorithm.

size

number of hidden units in the neural network.

lag

Number of lags to include in each regime.

n.ahead

The forecast horizon.

ngridBeta

number of elements to search for the cointegrating value.

ngridTh

number of elements to search for the threshold value.

th1

different possibilities to pre-specify an exact value, an interval or a central point for the search of the threshold.

th2

different possibilities to pre-specify an exact value or a central point for the search of the second threshold.

beta0

Additional regressors to include in the cointegrating relation.

See also

Other Learners: Custom_chain, Lrnr_HarmonicReg, Lrnr_arima, Lrnr_base, Lrnr_condensier, Lrnr_cv, Lrnr_define_interactions, Lrnr_expSmooth, Lrnr_rugarch, Lrnr_sl, Pipeline, Stack