xgboost: eXtreme Gradient Boosting Definition

Definition of xgboost type models. This function is for internal use only.

This learner provides fitting procedures for xgboost models, using the xgboost package. Such models are classification and regression trees with extreme gradient boosting. For details on the fitting procedure, consult the documentation of the xgboost package.

define_xgboost_X(task, add_outcome = FALSE)

Lrnr_xgboost

Arguments

task
An object of type Lrnr_base as defined in this package.

add_outcome
A logical indicating whether the trained model object should be fit in a supervised fashion (with the outcome vector) or in an unsupervised fashion.

Format

R6Class object.

Value

Lrnr_base object with methods for training and prediction.