Call with filtered argument list

Call a function with a list of arguments, eliminating any that aren't matched in the function prototype

call_with_args(fun, args, other_valid = list(), keep_all = FALSE)

Arguments

fun
A function whose signature will be used to reduce the

args
A list of function arguments to use.

other_valid
A list of function arguments names that are valid, but not formals of fun.

keep_all
A boolean don't drop arguments, even if they aren't matched in either the function prototype or other_valid.