Subset of growth data from the collaborative perinatal project (CPP)

Subset of growth data from the collaborative perinatal project (CPP).

cpp

Source

https://catalog.archives.gov/id/606622

Broman, Sarah. 'The collaborative perinatal project: an overview.' Handbook of longitudinal research 1 (1984): 185-227.

Examples

data(cpp)
head(cpp)

#>   subjid agedays   wtkg htcm lencm      bmi   waz   haz   whz   baz siteid sexn
#> 1      1       1  4.621   55    55 15.27603  2.38  2.61  0.19  1.35      5    1
#> 2      1     123  8.760   NA    NA       NA  1.99    NA    NA    NA      5    1
#>    sex feedingn feeding gagebrth birthwt birthlen apgar1 apgar5 mage mracen
#> 1 Male       90 Unknown      287    4621       55      8      9   21      5
#> 2 Male       90 Unknown      287    4621       55      8      9   21      5
#>   mrace mmaritn  mmarit meducyrs sesn          ses parity gravida smoked
#> 1 White       1 Married       12   50       Middle      1       1      0
#> 2 White       1 Married       12   50       Middle      1       1      0
#>   mcignum preeclmp comprisk geniq sysbp diabp
#> 1       0        0     none    NA    NA    NA
#> 2       0        0     none    NA    NA    NA
#>  [ reached getOption("max.print") -- omitted 4 rows ]