List sl3 Learners

Lists learners in sl3 (defined as objects that start with Lrnr_ and inherit from Lrnr_base)

sl3_list_properties()

sl3_list_learners(properties = c())

Arguments

properties
a vector of properties that learners must match to be returned

Value

a vector of learner names that match the property list