Make a stack of sl3 learners

Produce a stack of learners by passing in a list with IDs for the learners. The resultant stack of learners may then be used as normal.

make_learner_stack(...)

Arguments

...
Each argument is a list that will be passed to make_learner

Value

An sl3 Stack consisting of the learners passed in as arguments the list argument to this function. This Stack has all of the standard methods associated with such objects.

Examples

# constructing learners with default settings
sl_stack_easy <- make_learner_stack("Lrnr_mean", "Lrnr_glm_fast",
                                    "Lrnr_xgboost")

# constructing learners with arguments passed in
sl_stack <- make_learner_stack("Lrnr_mean",
                               list("Lrnr_condensier", nbins = 5,
                                    bin_method = "equal.len", pool = FALSE))