Drop components from learner fits

Given learner fit, sequentially drop components from internal fit object, keeping track of which components are needed for prediction.

reduce_fit_test(learner_fit)

Arguments

learner_fit

A Lrnr_* object, inheriting from Lrnr_base, that has been trained.