Generate a file containing a template sl3 Learner

Generates a template file that can be used to write a new sl3 Learner. For more information, see the Defining New Learners vignette.

write_learner_template(file)

Arguments

file
the path where the file should be written

Value

the return from file.copy. TRUE if writing the template was successful.