The Scalable Highly Adaptive Lasso

The Highly Adaptive Lasso is an estimation procedure that generates a design matrix consisting of basis functions corresponding to covariates and interactions of covariates and fits Lasso regression to this (usually) very wide matrix, recovering a nonparametric functional form that describes the target prediction function as a composition of subset functions with finite variation norm. This implementation uses hal9001, which provides both a custom implementation (based on origami) of the cross-validated lasso as well the standard call to cv.glmnet from the glmnet.

Format

R6Class object.

Value

Learner object with methods for training and prediction. See Lrnr_base for documentation on learners.

Parameters

max_degree=3

The highest order of interaction terms for which the basis functions ought to be generated. The default corresponds to generating basis functions up to all 3-way interactions of covariates in the input matrix, matching the default in hal9001.

fit_type="glmnet"

The specific routine to be called when fitting the Lasso regression in a cross-validated manner. Choosing the "glmnet" option calls either cv.glmnet or glmnet.

n_folds=10

Integer for the number of folds to be used when splitting the data for cross-validation. This defaults to 10 as this is the convention for V-fold cross-validation.

use_min=TRUE

Determines which lambda is selected from cv.glmnet. TRUE corresponds to "lambda.min" and FALSE corresponds to "lambda.1se".

reduce_basis=NULL

A numeric value bounded in the open interval (0,1) indicating the minimum proportion of ones in a basis function column needed for the basis function to be included in the procedure to fit the Lasso. Any basis functions with a lower proportion of 1's than the specified cutoff will be removed. This argument defaults to NULL, in which case all basis functions are used in the Lasso stage of HAL.

return_lasso=TRUE

A logical indicating whether or not to return the glmnet fit of the Lasso model.

return_x_basis=FALSE

A logical indicating whether or not to return the matrix of (possibly reduced) basis functions used in the HAL Lasso fit.

basis_list=NULL

The full set of basis functions generated from the input data (from enumerate_basis). The dimensionality of this structure is roughly (n * 2^(d - 1)), where n is the number of observations and d is the number of columns in the input.

cv_select=TRUE

A logical specifying whether the array of values specified should be passed to cv.glmnet in order to pick the optimal value (based on cross-validation) (when set to TRUE) or to fit along the sequence of values (or a single value using glmnet (when set to FALSE).

...

Other parameters passed directly to fit_hal. See its documentation for details.

See also