Nonlinear Time Series Analysis

This learner supports various forms of nonlinear autoregression, including additive AR, neural nets, SETAR and LSTAR models, threshold VAR and VECM.

Lrnr_tsDyn

Format

R6Class object.

Value

Learner object with methods for training and prediction. See Lrnr_base for documentation on learners.

Parameters

learner

Available built-in time series models. Currently available are listed by with availableModels() function.

m=1

embedding dimension.

size=1

number of hidden units in the neural network.

lag=1

Number of lags to include in each regime.

d=1

time delay.

include="const"

Type of deterministic regressors to include.

type="level"

Whether the variable is taken is level, difference or a mix as in the ADF test.

n.ahead="null"

The forecast horizon.

mL=m

autoregressive order for low regime.

mH=m

autoregressive order for high regime.

mM=NULL

autoregressive order for middle regime.

thDelay=0

Delay Time delay for the threshold variable.

common="none"

Indicates which elements are common to all regimes.

ML=seq_len(mL)

vector of lags for order for low.

MM=NULL

vector of lags for order for middle.

MH=seq_len(mH)

vector of lags for order for high.

nthresh=1

Threshold of the model.

trim=0.15

trimming parameter indicating the minimal percentage of observations in each regime.

sig=0.05

significance level for the tests to select the number of regimes.

control=list()

further arguments to be passed as control list to optim.

r=1

Number of cointegrating relationships.

model="VAR"

Model to estimate. Choices: VAR/VECM/TAR/MTAR.

I="level"

For VAR only: whether in the VAR variables are to be taken in levels or as a difference.

beta=NULL

For VECM only: imposed cointegrating value.

estim="2OLS"

Type of estimator for the VECM (two-step approach or Johansen MLE).

exogen=NULL

Inclusion of exogenous variables.

LRinclude="none"

Possibility to include in the long-run relationship and the ECT trend.

commonInter=FALSE

Whether the deterministic regressors are regime specific.

mTh=1

combination of variables with same lag order for the transition variable.

gamma=NULL

prespecified threshold values.

dummyToBothRegimes=TRUE

Whether the dummy in the one threshold model is applied to each regime.

max.iter=2

Number of iterations for the algorithm.

ngridBeta=50

Number of elements to search for the cointegrating value.

ngridTh=50

Number of elements to search for the threshold value.

th1

different possibilities to pre-specify an exact value, an interval or a central point for the search of the threshold.

th2

different possibilities to pre-specify an exact value or a central point for the search of the second threshold.

beta0=0

Additional regressors to include in the cointegrating relation.

...

Not currently used.

See also

Other Learners: Custom_chain, Lrnr_HarmonicReg, Lrnr_arima, Lrnr_base, Lrnr_bilstm, Lrnr_condensier, Lrnr_cv, Lrnr_define_interactions, Lrnr_expSmooth, Lrnr_glm_fast, Lrnr_glmnet, Lrnr_glm, Lrnr_h2o_grid, Lrnr_independent_binomial, Lrnr_lstm, Lrnr_mean, Lrnr_nnls, Lrnr_optim, Lrnr_pkg_SuperLearner, Lrnr_randomForest, Lrnr_rugarch, Lrnr_sl, Lrnr_solnp_density, Lrnr_solnp, Lrnr_subset_covariates, Lrnr_xgboost, Pipeline, Stack, define_h2o_X, undocumented_learner