Subset of growth data from the collaborative perinatal project (CPP)

Subset of growth data from the collaborative perinatal project (CPP). cpp_imputed drops observations for which the haz column is NA, and imputes all other observations as 0. This is only for the purposes of simplifying testing and examples.

data(cpp)

Format

A data frame with 1,912 repated-measures observations and 500 unique subjects:

subjid

Subject ID

agedays

Age since birth at examination (days)

wtkg

Weight (kg)

htcm

Standing height (cm)

lencm

Recumbent length (cm)

bmi

BMI (kg/m**2)

waz

Weight for age z-score

haz

Length/height for age z-score

whz

Weight for length/height z-score

baz

BMI for age z-score

siteid

Investigational Site ID

sexn

Sex (num)

sex

Sex

feedingn

Maternal breastfeeding status (num)

feeding

Maternal breastfeeding status

gagebrth

Gestational age at birth (days)

birthwt

Birth weight (gm)

birthlen

Birth length (cm)

apgar1

APGAR Score 1 min after birth

apgar5

APGAR Score 5 min after birth

mage

Maternal age at birth of child (yrs)

mracen

Maternal race (num)

mrace

Maternal race

mmaritn

Mothers marital status (num)

mmarit

Mothers marital status

meducyrs

Mother, years of education

sesn

Socio-economic status (num)

ses

Socio-economic status

parity

Maternal parity

gravida

Maternal num pregnancies

smoked

Maternal smoking status

mcignum

Num cigarettes mom smoked per day

comprisk

Maternal risk factors

Source

https://catalog.archives.gov/id/606622

Broman, Sarah. 'The collaborative perinatal project: an overview.' Handbook of longitudinal research 1 (1984): 185-227.

Examples

data(cpp)
head(cpp)

#>   subjid agedays   wtkg htcm lencm      bmi   waz   haz   whz   baz siteid sexn
#> 1      1       1  4.621   55    55 15.27603  2.38  2.61  0.19  1.35      5    1
#> 2      1     123  8.760   NA    NA       NA  1.99    NA    NA    NA      5    1
#> 3      1     366 14.500   79    79 23.23346  3.84  1.35  4.02  3.89      5    1
#> 4      2       1  3.345   51    51 12.86044  0.06  0.50 -0.64 -0.43      5    1
#> 5      2     123  4.340   NA    NA       NA -3.99    NA    NA    NA      5    1
#> 6      2     366  8.400   73    73 15.76281 -1.27 -1.17 -0.96 -0.80      5    1
#>    sex feedingn feeding gagebrth birthwt birthlen apgar1 apgar5 mage mracen
#> 1 Male       90 Unknown      287    4621       55      8      9   21      5
#> 2 Male       90 Unknown      287    4621       55      8      9   21      5
#> 3 Male       90 Unknown      287    4621       55      8      9   21      5
#> 4 Male       90 Unknown      280    3345       51      8      9   15      5
#> 5 Male       90 Unknown      280    3345       51      8      9   15      5
#> 6 Male       90 Unknown      280    3345       51      8      9   15      5
#>   mrace mmaritn  mmarit meducyrs sesn          ses parity gravida smoked
#> 1 White       1 Married       12   50       Middle      1       1      0
#> 2 White       1 Married       12   50       Middle      1       1      0
#> 3 White       1 Married       12   50       Middle      1       1      0
#> 4 White       1 Married       NA   NA            .      0       0      1
#> 5 White       1 Married       NA   NA            .      0       0      1
#> 6 White       1 Married       NA   NA            .      0       0      1
#>   mcignum comprisk
#> 1       0     none
#> 2       0     none
#> 3       0     none
#> 4      35     none
#> 5      35     none
#> 6      35     none

#