Subset of growth data from the collaborative perinatal project (CPP)

Subset of growth data from the collaborative perinatal project (CPP) at single time-point. The rows in original cpp data were subset for agedays==366. See ?cpp for the description of the variables.

data(cpp_1yr)

Source

https://catalog.archives.gov/id/606622

Broman, Sarah. 'The collaborative perinatal project: an overview.' Handbook of longitudinal research 1 (1984): 185-227.

Examples

data(cpp_1yr)
head(cpp_1yr)

#>    subjid agedays wtkg htcm lencm      bmi   waz   haz   whz   baz siteid sexn
#> 3       1     366 14.5   79    79 23.23346  3.84  1.35  4.02  3.89      5    1
#> 6       2     366  8.4   73    73 15.76281 -1.27 -1.17 -0.96 -0.80      5    1
#> 14      7     366 12.6   79    79 20.18907  2.46  1.35  2.38  2.21      5    1
#> 18      8     366 11.3   81    81 17.22298  1.45  2.20  0.72  0.31      5    1
#> 23      9     366 10.7   77    77 18.04689  1.43  1.15  1.28  1.10      5    2
#> 30     11     366  9.0   78    78 14.79290 -0.64  0.93 -1.39 -1.63      5    1
#>       sex feedingn                 feeding gagebrth birthwt birthlen apgar1
#> 3    Male       90                 Unknown      287    4621       55      8
#> 6    Male       90                 Unknown      280    3345       51      8
#> 14   Male       90                 Unknown      287    3260       53      6
#> 18   Male        2 Exclusively formula fed      287    3629       53      9
#> 23 Female        2 Exclusively formula fed      287    3657       55      8
#> 30   Male       90                 Unknown      266    2835       51      9
#>    apgar5 mage mracen mrace mmaritn  mmarit meducyrs sesn          ses parity
#> 3       9   21      5 White       1 Married       12   50       Middle      1
#> 6       9   15      5 White       1 Married       NA   NA            .      0
#> 14      7   19      5 White       1 Married       12   75        Upper      0
#> 18      9   22      5 White       1 Married       12   63 Upper-middle      1
#> 23      9   23      5 White       1 Married       12   63 Upper-middle      2
#> 30      9   27      3 Black       1 Married       NA   NA            .      0
#>    gravida smoked mcignum comprisk
#> 3        1      0       0     none
#> 6        0      1      35     none
#> 14       0      0       0     none
#> 18       1      0       0     none
#> 23       2      0       0     none
#> 30       0      0       0         

table(cpp_1yr[["agedays"]])

#> 
#> 366 
#> 428