Importance Extract variable importance measures produced by randomForest and order in decreasing order of importance.

Function that takes a cross-validated fit (i.e., cross-validated learner that has already been trained on a task), which could be a cross-validated single learner or super learner, to generate a risk-based variable importance measure for each covariate in the trained task, or for each group of covariates. This function outputs a data.table, where each row corresponds to the risk difference or the risk ratio between the following two risks: the risk when a covariate (or group of covariates) is permuted or removed, and the original risk (i.e., when all covariates are included as they were in the observed data). A higher risk ratio/difference corresponds to a more important covariate/group. A plot can be generated from the returned data.table by calling companion function plot_importance.

importance(fit, eval_fun = NULL, fold_number = "validation",
  type = c("remove", "permute"), importance_metric = c("ratio",
  "difference"), covariate_groups = NULL, B = 100L,
  cores = parallel::detectCores() - 1)

importance(fit, eval_fun = NULL, fold_number = "validation",
  type = c("remove", "permute"), importance_metric = c("ratio",
  "difference"), covariate_groups = NULL, B = 100L,
  cores = parallel::detectCores() - 1)

Arguments

fit
A trained cross-validated learner (e.g., cv stack, super learner), from which cross-validated predictions can be generated.

eval_fun
The evaluation function (risk or loss function) for evaluating the risk. Defaults vary based on the outcome type: squared error loss for continuous outcomes, and negative log-likelihood loss for discrete outcomes. See loss_functions and risk_functions for options.

fold_number
The fold number to use for obtaining the predictions from the fit. Either a positive integer for obtaining predictions from a specific fold's fit; "full" for obtaining predictions from a fit on all of the data, or "validation" (default) for obtaining cross-validated predictions, where the data used for training and prediction never overlaps across the folds. Note that if a positive integer or "full" is supplied, then there will be overlap between the data used for training and prediction.

type
Which method should be used to obscure the relationship between the covariate and the outcome. When type is "remove" (default), each covariate is removed from the task and the cross-validated learner is refit to this modified task and then predictions are obtained from this refit. When type is "permute", each covariate is permuted (sampled without replacement), and then predictions are obtained from this modified data with one permuted covariate.

importance_metric
Either "ratio" (default), which returns the risk with the permuted/removed X divided by observed risk with all X; or "difference", which returns the difference between the risk with the permuted/removed X and the observed risk.

covariate_groups
Optional named list covariate groups which will invoke variable importance evaluation at the group-level, by removing/permuting all covariates in the same group together. If covariates in the task are not specified in the list of groups, then those covariates will be added as additional single-covariate groups.

B
Optional integer specifying the number of permutations to average the permuted risk over when type = "permute".

cores
Number of cores for parallelizing the B permutations.

Value

A data.table of variable importance for each covariate.